requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
assemblyai==0.33.0
certifi==2024.8.30
Django==4.2.16
djangorestframework==3.15.2
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.9
pydantic==2.9.1
pydantic_core==2.23.3
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
websockets==13.0.1
